Move validate_synthetics_event into validators directory
